close_pos_session, pos_settlement_transaction_summary - isu fix (#99)

+++ org_init_scripts/214_pos_settlement_query.sql
@@ -41,7 +41,7 @@ begin
                a.particulars,
                min(a.date),
                max(a.date),
-               count(a.voucher_id),
+               count(a.voucher_id)::int,
                round(sum(a.bill_amount)::numeric, 2)::float
         from pos_counter_transaction a
         where a.settlement_id = $1
